+++ app/Http/Controllers/DirtRally/ChampionshipController.php
@@ -4,7 +4,7 @@ namespace App\Http\Controllers\DirtRally;
 
 use App\Http\Controllers\Controller;
 use App\Http\Requests\ChampionshipRequest;
-use App\Models\Championship;
+use App\Models\DirtRally\Championship;
 
 class ChampionshipController extends Controller
 {
