+++ app/Http/Controllers/DirtRally/TimesController.php
@@ -3,10 +3,10 @@
 namespace App\Http\Controllers\DirtRally;
 
 use App\Http\Controllers\Controller;
-use App\Models\Championship;
-use App\Models\Event;
-use App\Models\Season;
-use App\Models\Stage;
+use App\Models\DirtRally\Championship;
+use App\Models\DirtRally\Event;
+use App\Models\DirtRally\Season;
+use App\Models\DirtRally\Stage;
 
 use App\Http\Requests;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
