+++ app/Models/DirtRally/PointsSequence.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Models;
+namespace App\Models\DirtRally;
 
 class PointsSequence extends \Eloquent
 {
