+++ app/Services/Times.php
@@ -2,8 +2,8 @@
 
 namespace App\Services;
 
-use App\Models\Event;
-use App\Models\Season;
+use App\Models\DirtRally\Event;
+use App\Models\DirtRally\Season;
 use Illuminate\Database\Eloquent\Collection;
 
 class Times
