+++ resources/views/points-system/show.blade.php
@@ -8,13 +8,13 @@
 
 @section('content')
 
-    {!! Form::open(['route' => ['points-system.destroy', $system], 'method' => 'delete', 'class' => 'form-inline']) !!}
+    {!! Form::open(['route' => ['dirt-rally.points-system.destroy', $system], 'method' => 'delete', 'class' => 'form-inline']) !!}
     <a class="btn btn-small btn-warning"
-       href="{{ route('points-system.edit', $system) }}">Edit System</a>
+       href="{{ route('dirt-rally.points-system.edit', $system) }}">Edit System</a>
     {!! Form::submit('Delete System', array('class' => 'btn btn-danger')) !!}
     {!! Form::close() !!}
 
-    {!! Form::open(['route' => ['points-system.points', $system], 'class' => 'form-horizontal']) !!}
+    {!! Form::open(['route' => ['dirt-rally.points-system.points', $system], 'class' => 'form-horizontal']) !!}
 
     <table class="table">
         <thead>
